createIndex method fully implemented
